move election data loader
