+++ src/geoscore_de/data_flow/features/__init__.py
@@ -1,8 +1,8 @@
 from geoscore_de.data_flow.features.area import AreaFeature
 from geoscore_de.data_flow.features.base import BaseFeature, instantiate_feature
 from geoscore_de.data_flow.features.birth import BirthFeature
-from geoscore_de.data_flow.features.election_21 import Election21Feature
-from geoscore_de.data_flow.features.election_25 import Election25Feature
+from geoscore_de.data_flow.features.election.election_21 import Election21Feature
+from geoscore_de.data_flow.features.election.election_25 import Election25Feature
 from geoscore_de.data_flow.features.migration import MigrationFeature
 from geoscore_de.data_flow.features.osm_atm import OSMATMFeature
 from geoscore_de.data_flow.features.osm_hospitals import OSMHospitalsFeature
