+++ src/geoscore_de/data_flow/features/election/__init__.py
@@ -1,0 +1,3 @@
+from geoscore_de.data_flow.features.election.base import BaseElectionFeature
+
+__all__ = ["BaseElectionFeature"]
